requirements.txt
altgraph 
# @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
blinker==1.8.2
click==8.1.7
Faker==26.0.0
Flask==3.0.3
Flask-Cors==4.0.1
future 
# @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gunicorn==22.0.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
macholib 
# @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.5
packaging==24.1
python-dateutil==2.9.0.post0
six 
# @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
Werkzeug==3.0.3
zipp==3.19.2
python-dotenv
flask-pymongo
PyJWT
passlib
uuid